# openapi.yaml
openapi: 3.0.3
info:
  title: Runs API - Phase 4
  version: v1.0.0
  description: API for managing Runs, RSVPs, Attendance, and Photo Uploads.
servers:
  - url: /api # Base path for all API endpoints
    description: Development server
components:
  securitySchemes:
    cookieAuth:
      type: apiKey
      in: cookie
      name: next-auth.session-token # Adjust if your session cookie name is different
  schemas:
    ErrorResponse:
      type: object
      properties:
        message:
          type: string
        errors: # Optional, for validation errors
          type: object
          additionalProperties:
            type: array
            items:
              type: string
      required:
        - message
    User:
      type: object
      properties:
        id:
          type: string
          format: cuid
        name:
          type: string
          nullable: true
        email:
          type: string
          format: email
          nullable: true # Depending on NextAuth setup, email might be null for some users
        image:
          type: string
          format: url
          nullable: true
        role:
          type: string
          enum: [user, organizer, admin]
          default: user
...
    Attendance:
      type: object
      properties:
        id:
          type: string
          format: cuid
        runId:
          type: string
          format: cuid
        userId:
          type: string
          format: cuid
        markedAt:
          type: string
          format: date-time
          description: Timestamp when attendance was marked.
    AttendanceWithUser:
      allOf:
        - $ref: '#/components/schemas/Attendance'
        - type: object
          properties:
            user:
              $ref: '#/components/schemas/User' # Simplified user for attendance list
    Photo:
      type: object
      properties:
        id:
          type: string
          format: cuid
        storageKey:
          type: string
          description: Key of the photo in S3.
        url:
          type: string
          format: url
          nullable: true
          description: Full URL of the photo after confirmation.
        caption:
          type: string
          maxLength: 500
          nullable: true
        runId:
          type: string
          format: cuid
        uploaderId:
          type: string
          format: cuid
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
    PhotoDetails: # Photo details as returned in RunDetailed
      allOf:
        - $ref: '#/components/schemas/Photo'
        - type: object
          properties:
            uploadedBy:
              $ref: '#/components/schemas/User' # Simplified user for photo uploader
...
security:
  - cookieAuth: [] # Apply cookieAuth globally, operations can override if public
paths:
...
  /runs/{id}/attendance:
    parameters:
      - name: id
        in: path
        required: true
        description: ID of the run to mark attendance for
        schema:
          type: string
          format: cuid
    post:
      summary: Mark a user as attended for a run
      tags:
        - Attendance
      security:
        - cookieAuth: [] # Requires organizer/admin role
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                userId:
                  type: string
                  format: cuid
                  description: ID of the user to mark as attended.
              required:
                - userId
      responses:
        '201':
          description: Attendance marked successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AttendanceWithUser'
        '400':
          description: Invalid input or run ID format
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: Forbidden (insufficient permissions)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Run or User to mark not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '409':
          description: User already marked as attended for this run
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Internal Server Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /runs/{id}/photos/generate-signed-url:
    parameters:
      - name: id
        in: path
        required: true
        description: ID of the run to associate the photo with
        schema:
          type: string
          format: cuid
    post:
      summary: Generate a pre-signed URL for photo upload
      tags:
        - Photos
      security:
        - cookieAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                fileName:
                  type: string
                contentType:
                  type: string
                  example: image/jpeg
              required:
                - fileName
                - contentType
      responses:
        '200':
          description: Signed URL generated successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  signedUrl:
                    type: string
                    format: url
                  photoId:
                    type: string
                    format: cuid
                  storageKey:
                    type: string
        '400':
          description: Invalid input or run ID format
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Run not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Internal Server Error or S3 configuration issue
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /runs/{id}/photos/confirm-upload:
    parameters:
      - name: id
        in: path
        required: true
        description: ID of the run (for route namespacing)
        schema:
          type: string
          format: cuid
    post:
      summary: Confirm photo upload and save metadata
      tags:
        - Photos
      security:
        - cookieAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                photoId:
                  type: string
                  format: cuid
                caption:
                  type: string
                  maxLength: 500
                  nullable: true
              required:
                - photoId
      responses:
        '200':
          description: Photo upload confirmed and metadata saved
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PhotoDetails'
        '400':
          description: Invalid input
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Photo record not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Internal Server Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'